+++ apis/events.d.ts
@@ -55,7 +55,7 @@ export class Request<T = any> extends Event<T> {
 
   res: express.Response
 
-  params: (string | object)[]
+  params: Record<string, any>[]
 
   method: string
 
@@ -74,6 +74,8 @@ export class Request<T = any> extends Event<T> {
   subject: ref
 
   reply (results: any): void
+  /** @beta */
+  reply (results: any, options: { mimetype?: string, filename?: string, [key: string]: any }): void
 
   notify (code: number, message: string, target?: string, args?: any[]): Error
 
@@ -83,7 +85,7 @@ export class Request<T = any> extends Event<T> {
 
   error (code: number, message: string, target?: string, args?: any[]): Error
 
-  reject (code: number, message: string, target?: string, args?: any[]): Error
+  reject (code: number, message: string, target?: string, args?: any[]): never
 
   notify (code: number, message: string, args?: any[]): Error
 
@@ -93,7 +95,7 @@ export class Request<T = any> extends Event<T> {
 
   error (code: number, message: string, args?: any[]): Error
 
-  reject (code: number, message: string, args?: any[]): Error
+  reject (code: number, message: string, args?: any[]): never
 
   notify (message: string, target?: string, args?: any[]): Error
 
@@ -103,7 +105,7 @@ export class Request<T = any> extends Event<T> {
 
   error (message: string, target?: string, args?: any[]): Error
 
-  reject (message: string, target?: string, args?: any[]): Error
+  reject (message: string, target?: string, args?: any[]): never
 
   notify (message: { code?: number | string, message: string, target?: string, args?: any[] }): Error
 
@@ -113,7 +115,7 @@ export class Request<T = any> extends Event<T> {
 
   error (message: { code?: number | string, message: string, target?: string, args?: any[], status?: number }): Error
 
-  reject (message: { code?: number | string, message: string, target?: string, args?: any[], status?: number }): Error
+  reject (message: { code?: number | string, message: string, target?: string, args?: any[], status?: number }): never
 
 }