+++ apis/facade.d.ts
@@ -7,6 +7,7 @@ export * from './events'
 export * from './utils'
 export * from './cqn'
 export * from './global'
+export * from './i18n'
 export { log, debug } from './log'
 export { test } from './test'
 
@@ -44,3 +45,5 @@ export const cli: {
 
 import { env } from './env'
 export const requires: env.Requires
+
+export * from '@sap/cds-dk'  // these only contribute types if the user has cds-dk installed