+++ CHANGELOG.md
@@ -5,10 +5,47 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
+
+### Added
+- Added type export for types from cds-dk. These types are only available if `@sap/cds-dk` is also present in the project.
+- Optional parameter 'options' in `req.reply`. The object can contain, for example, mimetype and/or filename.
+- Added types for `req.req` and `req.res` in `cds.Request`
+### Changed
+- [breaking] Bump dependency to `@sap/cds` to `>=9.0.0`. This means starting with this version, `cds-types` is supposed to be used alongside `@sap/cds@9`!
+- `req.params` always returns an array of objects
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [0.11.0] - 2025-06-05
+
+### Added
+- Syntactically allow infix filters in `SELECT` clauses through tagged templates. These filters are not semantically checked for validity
+- Added `cds.linked.LinkedDefinition` as alias for `cds.linked.classes.any_`
+- Added `doc?: string` to `cds.linked.classes.any_`
+- Add overload for `Service.emit` to offer improved type support when using an event type emitted by cds-typer.
+- Added programmatic draft actions.
+- Added types for `util.path`, `util.fs`, and `util.inspect`.
+
+### Changed
+- [breaking] Corrected the way the default export is generated. This also gets rid of the export `default_2` that was mistakenly exposed before.
+- Renamed `CQN` property `INSERT.as` to `INSERT.from`
+- `Request.reject(…)` now returns `never` instead of `Error`, as its implementation always throws.
+
+### Deprecated
+### Removed
+- Deprecated `INSERT.from` method of `cds.ql` API
+
+### Fixed
+### Security
+
+## [0.10.0] - 2025-04-01
 ### Added
 - Added support for new builtin type `cds.Map`
 - Added types for `SELECT.hints()` of `cds.ql` API
-- Added types for `req.req` and `req.res` in `cds.Request`
+- Added types for `.bind(Service)` to all queries.
+- Added types for `i18n` module
 
 ### Changed
 - `CHANGELOG.md` and `LICENSE` files are no longer part of the npm package.